prevent ppl from deleting non-active job

+++ src/pages/api/jobs/[job_id].js
@@ -1,8 +1,13 @@
-import { ErrorHandler } from '../../../utils/ErrorHandler'
+import {
+	ErrorHandler,
+	HttpBadRequestError,
+} from '../../../utils/ErrorHandler'
 import { ApiHelper } from '../../../utils/ApiHelper'
 import {
+	ERROR,
 	HTTP_METHOD,
 	HTTP_STATUS,
+	JOB_STATUS,
 } from '../../../utils/constants'
 import { JobDAO } from '../../../dao/JobDAO'
 
@@ -16,6 +21,10 @@ export default async function handler(req, res) {
 
 		if (req.method === HTTP_METHOD.DELETE) {
 			await ApiHelper.checkUser(req)
+			const job = await JobDAO.getOneById(job_id, ['status'])
+			if (job.status !== JOB_STATUS.ACTIVE) {
+				throw new HttpBadRequestError(ERROR.CANNOT_DELETE_NON_ACTIVE_JOB)
+			}
 			await JobDAO.delete(job_id)
 			res.status(HTTP_STATUS.OK).end()
 		}
